// panel/web/src/lib/types.ts
export type BlockType = 'hero' | 'download' | 'servers' | 'leaderboard' | 'stats' | 'instances' | 'news' | 'faq' | 'socials' | 'text' | 'image' | 'split' | 'features' | 'gallery' | 'cta' | 'divider' | 'testimonials' | 'video' | 'buttons' | 'map';

export interface LandingBlock {
  id: string;
  type: BlockType;
  enabled: boolean;
  title?: string;
  subtitle?: string;
  options?: Record<string, any>;
}